Restructure and clean up repeated code
