+++ admin-api-frontend/src/components/FormsFolder/AddRolesForm/index.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+import CommonFormWithRoles from "../../CommonFormsFolder/CommonFormWithRoles/index.tsx";
+
+const AddRolesForm = () => {
+  return (
+    <CommonFormWithRoles name="AddRolesForm" backendFunction={console.log} />
+  );
+};
+export default AddRolesForm;
