+++ admin-api-frontend/src/components/FormsFolder/CreateUserForm/index.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+import CommonFormWithRoles from "../../CommonFormsFolder/CommonFormWithRoles/index.tsx";
+
+const CreateUserForm = () => {
+  return (
+    <CommonFormWithRoles name="CreateUserForm" backendFunction={console.log} />
+  );
+};
+export default CreateUserForm;
