+++ admin-api-frontend/src/components/FormsFolder/DeleteUserForm/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+import CommonForm from "../../CommonFormsFolder/CommonForm/index.tsx";
+
+const DeleteUserForm = () => {
+  return <CommonForm name="DeleteUserForm" backendFunction={console.log} />;
+};
+export default DeleteUserForm;
