+++ admin-api-frontend/src/components/FormsFolder/GetUserInfoForm/index.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+
+import CommonForm from "../../CommonFormsFolder/CommonForm/index.tsx";
+
+const GetUserInfoForm = () => {
+  return <CommonForm name="GetUserInfoForm" backendFunction={console.log} />;
+};
+export default GetUserInfoForm;
