+++ admin-api-frontend/src/components/FormsFolder/RemoveRolesForm/index.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+import CommonFormWithRoles from "../../CommonFormsFolder/CommonFormWithRoles/index.tsx";
+
+const RemoveRolesForm = () => {
+  return (
+    <CommonFormWithRoles name="RemoveRolesForm" backendFunction={console.log} />
+  );
+};
+export default RemoveRolesForm;
